+++ tests/register.test.ts
@@ -1,7 +1,7 @@
 import test from 'tape'
 
 import { charlieStashSeed, setupTest } from './testing-utils'
-import { register } from 'src/flows/register/register'
+import { register } from '../src/flows/register/register'
 import { readFileSync } from 'node:fs'
 
 const networkType = 'two-nodes'
@@ -13,7 +13,7 @@ test('Regsiter - Default Program', async (t) => {
 
   const fullAccount = entropy.keyring.getAccount()
 
-  t.equal(verifyingKey, fullAccount.registration.verifyingKeys[0], 'verifying key matches key added to registration account')
+  t.equal(verifyingKey, fullAccount?.registration?.verifyingKeys?.[0], 'verifying key matches key added to registration account')
 
   t.end()
 })
@@ -21,7 +21,7 @@ test('Regsiter - Default Program', async (t) => {
 test('Register - Barebones Program', async t => {
   const { run, entropy } = await setupTest(t, { networkType, seed: charlieStashSeed })
   const dummyProgram: any = readFileSync(
-    'src/programs/template_barebones.wasm'
+    new URL('./programs/template_barebones.wasm', import.meta.url)
   )
   const pointer = await run(
     'deploy program',
@@ -37,8 +37,8 @@ test('Register - Barebones Program', async t => {
   )
 
   const fullAccount = entropy.keyring.getAccount()
-  
-  t.equal(verifyingKey, fullAccount.registration.verifyingKeys[1], 'verifying key matches key added to registration account')
+
+  t.equal(verifyingKey, fullAccount?.registration?.verifyingKeys?.[1], 'verifying key matches key added to registration account')
 
   t.end()
 })